requirements.txt
bittensor>=6.0.0
pydantic>=2.0
numpy>=1.24.0

# Platform API
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
pydantic[email]>=2.0
httpx>=0.24.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
